+++ operator/controllers/http/httpscaledobject_controller.go
@@ -120,7 +120,7 @@ func (r *HTTPScaledObjectReconciler) Reconcile(ctx context.Context, req ctrl.Req
 	)
 
 	// Create required app objects for the application defined by the CRD
-	if err := createOrUpdateApplicationResources(
+	if err := r.createOrUpdateApplicationResources(
 		ctx,
 		logger,
 		r.Client,